+++ typescript/packages/management/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.4.1](https://github.com/olane-labs/copass/compare/@copass/management@0.1.2...@copass/management@0.4.1) (2026-05-02)
+
+### Bug Fixes
+
+- strip JSON-Schema meta keys from spec corpus + auto-sync python wheel ([cbb7021](https://github.com/olane-labs/copass/commit/cbb7021b75f674b7fe57a3b10cd5fd3177046678))
+
 ## [0.1.2](https://github.com/olane-labs/copass-harness/compare/@copass/management@0.1.1...@copass/management@0.1.2) (2026-05-01)
 
 **Note:** Version bump only for package @copass/management